+++ app/src/components/search/filters/selectFilter.tsx
@@ -152,14 +152,29 @@ const SelectFilterComponent = ({
           accordionButtonRef.current?.focus();
           searchManager.setLastFilter(`accordion-button-select-${filter.name}`);
           // Push the current filter selection to URL.
-          updateURL(
-            searchManager.handleAddFilter([
-              {
-                filter: filter.name,
-                value: current?.name!,
-              },
-            ])
-          );
+          if (filter.name === "collection") {
+            const [collectionTitle, collectionUuid] =
+              current?.name!.split("||") ?? [];
+            updateURL(
+              searchManager.handleAddFilter([
+                {
+                  filter: filter.name,
+                  value: `${encodeURIComponent(
+                    collectionTitle
+                  )}||${encodeURIComponent(collectionUuid)}`,
+                },
+              ])
+            );
+          } else {
+            updateURL(
+              searchManager.handleAddFilter([
+                {
+                  filter: filter.name,
+                  value: current?.name!,
+                },
+              ])
+            );
+          }
           setUserClickedOutside(true);
         }}
       >